Fix installation problems (package data and conf file)
The package did not contain package data, which triggered an error when
installing the package with pip because .rst files are not found. Also,
if the user never ran `eda configure` before using other functions
requiring configuration data, the user configuration file does not exist
yet, which triggers an error.

We fix both of these problems in this commit.